redo project structure
