+++ packages/rift/src/components/buttons/CallActions.js
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { useSelector } from 'react-redux';
-import IncomingCall from '../IncomingCall';
+import IncomingCall from '../talk/IncomingCall';
 
 import {
 	Layout,
