+++ packages/rift/src/components/users/SearchBar.js
@@ -3,11 +3,11 @@ import { StyleSheet } from 'react-native';
 
 import { Layout, Text, Button, styled, Input } from '@ui-kitten/components';
 import { connect } from 'react-redux';
-import * as Actions from '../actions';
-import * as Selectors from '../selectors';
-import * as CallSelectors from '../selectors/call';
-import * as UserSelectors from '../selectors/users';
-import * as AuthSelectors from '../selectors/auth';
+import * as Actions from '@actions';
+import * as Selectors from '@selectors';
+import * as CallSelectors from '@selectors/call';
+import * as UserSelectors from '@selectors/users';
+import * as AuthSelectors from '@selectors/auth';
 // export const SearchBar = (props) => {
 //     return ()
 // }
