+++ packages/rift/src/components/users/SearchBar.stories.js
@@ -1,17 +1,12 @@
 import * as React from 'react';
 // import { action } from '@storybook/addon-actions';
-import { Media } from './Media';
+import SearchBar from './SearchBar';
 // import { taskData, actionsData } from './Task.stories';
-import Provider from '../storybook/Provider';
-import createStore from '../store';
+import Provider from '../../storybook/Provider';
+import createStore from '../../store';
 const store = createStore({});
 export const defaultSearchData = {
-	audioRef: React.createRef(),
-	videoRef: React.createRef(),
-	callFunctions: {
-		audio: () => true,
-		video: () => true
-	}
+	filterText: 'hey'
 };
 
 // export const actionsData = {
@@ -21,14 +16,19 @@ export const defaultSearchData = {
 
 const withProvider = story => <Provider store={store}>{story()}</Provider>;
 export default {
-	component: Media,
-	title: 'Media',
+	component: SearchBar,
+	title: 'SearchBar',
 	//   decorators: [story => <div style={{ padding: '3rem' }}>{story()}</div>],
 	decorators: [
 		withProvider,
 		story => (
 			<div
-				style={{ display: 'flex', backgroundColor: '#222C44', padding: '3rem' }}
+				style={{
+					height: '200px',
+					width: '400px',
+					backgroundColor: '#222C44',
+					padding: '3rem'
+				}}
 			>
 				{story()}
 			</div>
@@ -37,9 +37,4 @@ export default {
 	excludeStories: /.*Data$/
 };
 
-export const Default = () => (
-	<Media
-		tasks={defaultSearchData}
-		callFunctions={defaultSearchData.callFunctions}
-	/>
-);
+export const Default = () => <SearchBar tasks={defaultSearchData} />;
