+++ packages/rift/src/components/users/TabBar.js
@@ -1,3 +1,4 @@
+// REMOVE
 import React, { useState } from 'react';
 import {
 	Layout,
