+++ packages/rift/src/containers/misc/About.js
